DTSGen.py: Use a longer RNG seed for XiangShan NEMU board
Currently, the RNG seed used in the XiangShan NEMU board is
"NEMU_BOARD", which is only 10 bytes long. This short seed may not
provide sufficient entropy for kernel crng initialization, leading to
user applications that rely on /dev/urandom to block during startup.

To address this issue, we propose to extend the RNG seed to
"XIANGSHAN_NEMU_BOARD_RANDOM_SEED", which is 32 bytes long. This longer
seed will provide better entropy, reducing the likelihood of blocking
during application startup.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>